+++ MovieQuiz/Models/QuizReslultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizReslultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Mikhail Kolokolnikov on 01.12.2022.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
